requirements.txt
aiohttp==3.7.2
aiojira==0.1.14
asyncio==3.4.3
django-cors-headers==3.6.0
django==3.0.5
djangorestframework==3.12.2
flake8==3.8.4
jira==2.0.0
p4python==2020.1.1983437
pandas==1.1.4
psycopg2-binary==2.8.6
PyJWT==1.7.1
PyMySQL==1.0.2
pytest==6.1.2
sqlparse==0.2.4
uWSGI>=2.0.18,<2.1
xlrd==1.2.0